V0.4.4 RC1
Release candidate
fixes to time stamping in the L2lidar class (already released as V1.1)
corrected aggregation default from 39 to 38
change limit on point cloud buffer size to 8 million .

+++ L2lidarClass/include/PCpoint.h
@@ -10,7 +10,7 @@ struct PCpoint
     float z;
     float intensity; //  0-255
     float range;    // meters
-    float time; // timestamp in seconds
+    long long time; // timestamp in nanoseconds
     uint32_t ring;
 };
